feat: create `TraySpy` struct for wrapping spy window logic
