+++ crates/systray-util/src/lib.rs
@@ -1,10 +1,12 @@
 mod error;
 mod tray;
+mod tray_spy;
 mod util;
 mod window;
 
 pub use error::*;
 pub use tray::*;
+pub use tray_spy::*;
 pub use util::*;
 pub use window::*;
